+++ src-core/src/main.rs
@@ -238,6 +238,7 @@ async fn app_setup(app_handle: tauri::AppHandle) {
             error!("[Core] Failed to register deep link schemas: {e}");
         }
     }
+    os::register_notification_app_id(&app_handle);
     // Initialize utility module
     utils::init();
     // Initialize Steam module